+++ ft_strmapi.c
@@ -6,13 +6,13 @@
 /*   By: nouillebobby <nouillebobby@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 00:14:09 by lcluzan           #+#    #+#             */
-/*   Updated: 2024/05/30 11:06:49 by nouillebobb      ###   ########.fr       */
+/*   Updated: 2024/05/30 11:37:50 by nouillebobb      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strmapi(char const *s, char (*f)(size_t, char))
+char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	size_t	i;
 	char	*new;
